+++ src/components/common/inputs/RadioButton.tsx
@@ -1,16 +1,13 @@
+import { ReactElement, ReactNode } from 'react';
 import { Radio } from '@chakra-ui/react';
 
 interface Props {
 	value?: string;
-	children: React.ReactNode;
+	children: ReactNode;
 	[prop: string]: any;
 }
 
-export default function RadioButton({
-	value,
-	children,
-	...props
-}: Props): React.ReactElement {
+export default function RadioButton({ value, children, ...props }: Props): ReactElement {
 	return (
 		<Radio value={value} variant='buttonStyle' position='relative' {...props}>
 			{children}